Move files to be processed with webpack outside static directory
