+++ src/html/modal.html
@@ -4,64 +4,64 @@
     <p class='modal-text symbol-text'> %s </p>
     <div class='modal-text'>
       <p class='name-text data'> %s </p>
-      <img class="copy" src="copy_icon_white.png"/>
+      <img class="copy" src="../res/copy_icon_white.png"/>
     </div>
     <p class='modal-text'> %s </p>
   </div>
   <div class='modal-body'>
     <div class='modal-text'>
       <p> Group </p> <p class="data"> : %s </p>
-      <img class="copy" src="copy_icon_black.png"/>
-      <img class="help tooltipcontainer" src="help_icon_black.png"/>
+      <img class="copy" src="../res/copy_icon_black.png"/>
+      <img class="help tooltipcontainer" src="../res/help_icon_black.png"/>
       <p class="tooltip">The periodic table can be divided into groups, or columns.</p>
     </div>
 
     <div class='modal-text'>
       <p> Period </p> <p class="data"> : %s </p>
-      <img class="copy" src="copy_icon_black.png"/>
-      <img class="help tooltipcontainer" src="help_icon_black.png"/>
+      <img class="copy" src="../res/copy_icon_black.png"/>
+      <img class="help tooltipcontainer" src="../res/help_icon_black.png"/>
       <p class="tooltip">The periodic table can be divided into periods, or rows.</p>
     </div>
 
     <div class='modal-text'>
       <p> Atomic Number </p> <p class="data"> : %s </p>
-      <img class="copy" src="copy_icon_black.png"/>
-      <img class="help tooltipcontainer" src="help_icon_black.png"/>
+      <img class="copy" src="../res/copy_icon_black.png"/>
+      <img class="help tooltipcontainer" src="../res/help_icon_black.png"/>
       <p class="tooltip">An element's atomic number is the number of protons an atom of that element has.</p>
     </div>
 
     <div class='modal-text'>
       <p> Atomic Mass </p> <p class="data"> : %s </p>
-      <img class="copy" src="copy_icon_black.png"/>
-      <img class="help tooltipcontainer" src="help_icon_black.png"/>
+      <img class="copy" src="../res/copy_icon_black.png"/>
+      <img class="help tooltipcontainer" src="../res/help_icon_black.png"/>
       <p class="tooltip">The atomic mass (or relative atomic mass) of an element is the weighted mean mass of an atom of that element compared with 1/12 the mass of an atom of Carbon-12.</p>
     </div>
 
     <div class='modal-text tag-text %s'>
       <p> Block </p> <p class="data"> : %s </p>
-      <img class="copy" src="copy_icon_white.png"/>
-      <img class="help tooltipcontainer" src="help_icon_white.png"/>
+      <img class="copy" src="../res/copy_icon_white.png"/>
+      <img class="help tooltipcontainer" src="../res/help_icon_white.png"/>
       <p class="tooltip">A block is a set of groups wherein the highest-energy electrons in each element belong to the same orbital type (the 's', 'd, 'p' and 'f' orbitals).</p>
     </div>
 
     <div class='modal-text tag-text %s'>
       <p> Type </p> <p class="data"> : %s </p>
-      <img class="copy" src="copy_icon_white.png"/>
-      <img class="help tooltipcontainer" src="help_icon_white.png"/>
+      <img class="copy" src="../res/copy_icon_white.png"/>
+      <img class="help tooltipcontainer" src="../res/help_icon_white.png"/>
       <p class="tooltip">Elements can also be grouped by type. Elements of the same type are often in the same or a similar group, and have very similar chemical properties.</p>
     </div>
 
     <div class='modal-text tag-text %s'>
       <p> State at room temperature </p> <p class="data"> : %s </p>
-      <img class="copy" src="copy_icon_white.png"/>
-      <img class="help tooltipcontainer" src="help_icon_white.png"/>
+      <img class="copy" src="../res/copy_icon_white.png"/>
+      <img class="help tooltipcontainer" src="../res/help_icon_white.png"/>
       <p class="tooltip">Elements all have different melting and boiling points, so not all are the same state at room temperature.</p>
     </div>
 
     <div class='modal-text tag-text %s'>
       <p> Electronegativity </p> <p class="data"> : %s </p>
-      <img class="copy" src="copy_icon_white.png"/>
-      <img class="help tooltipcontainer" src="help_icon_white.png"/>
+      <img class="copy" src="../res/copy_icon_white.png"/>
+      <img class="help tooltipcontainer" src="../res/help_icon_white.png"/>
       <p class="tooltip">The electronegativity of an atom is its ability to attract the shared pair of electrons in a covalent bond. Here it is measure using the Pauling scale.</p>
     </div>
   </div>
